requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.1
coreapi-cli==1.0.6
coreschema==0.0.4
decorator==4.1.2
django==1.9.1
djangorestframework==3.6.4
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.0
Jinja2==2.9.6
MarkupSafe==1.0
parso==0.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
win-unicode-console==0.5
